pyproject.toml
[project]
name = "Model_Context_Protocol"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.51.0",
    "arxiv>=2.2.0",
    "mcp>=1.9.0",
    "nest-asyncio>=1.6.0",
    "python-dotenv>=1.1.0",
]

[dependency-groups]
dev = [
    "arxiv>=2.2.0",
    "mcp>=1.9.0",
]
